Create Question.txt
the object
match the request. If at least one parameter doesn’t match, returns always False.
• find_matches: the method gets an array of objects of the class RealtyObject and checks all
the objects one by one. The method adds an object to the matching_objects attribute if this
object matches (use the previous method to check this). If check_match raises an exception,
add the corresponding realty object to the possible_objects list.
Derive two classes from the Request class: Request_apartment and Request_house.
Request_apartment has two additional attributes:
• floor_min (integer).
• floor_max (integer).
and the matching_objects and possible_objects attributes now include objects of the class
Apartment.
Rewrite the method check_match such that it checks in addition if the floor attribute of the corresponding
apartment is in the interval [floor_min, floor_max].
Request_house has one additional attribute: plotarea_min (double value): the minimal value of
the private land that the customer wants to get. The matching_objects and possible_objects
attributes now include objects of the class House.
Rewrite the method check_match such that it checks in addition if the plotarea attribute of the
corresponding house is larger or equal to plotarea_min of the request.
Add a main program to test your classes:
• Create the empty lists for the Apartment and House objects.
• Read in data from the Apartments and Houses files (the order of fields corresponds to the
structure of the class attributes).
• Do the same with the requests (use files Requests_apartments and Requests_houses).
• For each of the requests find all the matching objects.
• Print the matching_objects and possible_objects arrays for every request.
• Check the change_price method and print the result for some of the objects.